refactor: List Data 관련 컴포넌트 훅 정의
- List 컴포넌트에서 관리하던 상태를 page 컴포넌트에서 관리하는 것으로 변경
- 상태 관리를 커스텀 훅으로 정리
