+++ src/manager/hooks/campus/index.ts
@@ -1,0 +1,1 @@
+export * from './useCampusListData';
