+++ src/manager/hooks/course/index.ts
@@ -1,0 +1,1 @@
+export * from './useCourseListData';
