+++ src/manager/hooks/feed/index.ts
@@ -1,0 +1,1 @@
+export * from './useFeedListData';
