+++ src/manager/hooks/runningmate/index.ts
@@ -1,0 +1,1 @@
+export * from './useRunningmateListData';
