--- src/manager/pages/Course/CourseList.jsx
@@ -1,6 +1,0 @@
-import CourseList from '../../components/course/list/CourseList';
-const CourseListPage = () => {
-  return <CourseList />;
-};
-
-export default CourseListPage;
