--- src/manager/pages/Runningmate/RunningmateList.jsx
@@ -1,6 +1,0 @@
-import RunningmateList from '../../components/runningmate/list/RunningmateList';
-const RunningmateListPage = () => {
-  return <RunningmateList />;
-};
-
-export default RunningmateListPage;
