--- src/manager/pages/User/UserList.jsx
@@ -1,6 +1,0 @@
-import UserList from '../../components/user/list/UserList';
-const UserListPage = () => {
-  return <UserList />;
-};
-
-export default UserListPage;
